+++ src/FlaUILibrary/keywords/combobox.py
@@ -1,7 +1,7 @@
 from robotlibcore import keyword
 from FlaUILibrary.flaui.enum import InterfaceType
 from FlaUILibrary.flaui.module import (Combobox, Selector)
-from FlaUILibrary.flaui.automation.uia import UIA
+from FlaUILibrary.flaui.util.automationinterfacecontainer import AutomationInterfaceContainer
 
 
 class ComboBoxKeywords:
@@ -9,13 +9,13 @@ class ComboBoxKeywords:
     Interface implementation from robotframework usage for combobox keywords.
     """
 
-    def __init__(self, module: UIA):
+    def __init__(self, container: AutomationInterfaceContainer):
         """
         Constructor for combobox keywords.
 
-        ``module`` Automation framework module like UIA3 to handle element interaction.
+        ``container`` User automation container to handle element interaction.
         """
-        self._module = module
+        self._container = container
 
     @keyword
     def get_all_names_from_combobox(self, identifier, msg=None):
@@ -31,10 +31,11 @@ class ComboBoxKeywords:
         Examples:
         | ${data}  Get All Names From Combobox  <XPATH> <MSG>                   |
         """
-        element = self._module.get_element(identifier, InterfaceType.COMBOBOX, msg)
-        return self._module.action(Selector.Action.GET_ALL_NAMES,
-                                   Selector.create_value_container(element=element, msg=msg),
-                                   msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.COMBOBOX, msg)
+        return module.action(Selector.Action.GET_ALL_NAMES,
+                             Selector.create_value_container(element=element, msg=msg),
+                             msg)
 
     @keyword
     def get_all_texts_from_combobox(self, identifier, msg=None):
@@ -50,10 +51,11 @@ class ComboBoxKeywords:
         Examples:
         | ${data}  Get All Texts From Combobox  <XPATH> <MSG>                   |
         """
-        element = self._module.get_element(identifier, InterfaceType.COMBOBOX, msg)
-        return self._module.action(Selector.Action.GET_ALL_TEXTS,
-                                   Selector.create_value_container(element=element, msg=msg),
-                                   msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.COMBOBOX, msg)
+        return module.action(Selector.Action.GET_ALL_TEXTS,
+                             Selector.create_value_container(element=element, msg=msg),
+                             msg)
 
     @keyword
     def get_all_selected_texts_from_combobox(self, identifier, msg=None):
@@ -73,10 +75,11 @@ class ComboBoxKeywords:
         Examples:
         | ${data}  Get All Selected Texts From Combobox  <XPath>   |
         """
-        element = self._module.get_element(identifier, InterfaceType.COMBOBOX, msg)
-        return self._module.action(Selector.Action.GET_ALL_TEXTS_FROM_SELECTION,
-                                   Selector.create_value_container(element=element, msg=msg),
-                                   msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.COMBOBOX, msg)
+        return module.action(Selector.Action.GET_ALL_TEXTS_FROM_SELECTION,
+                             Selector.create_value_container(element=element, msg=msg),
+                             msg)
 
     @keyword
     def get_all_selected_names_from_combobox(self, identifier, msg=None):
@@ -96,10 +99,11 @@ class ComboBoxKeywords:
         Examples:
         | ${data}  Get All Selected Names From Combobox  <XPath>   |
         """
-        element = self._module.get_element(identifier, InterfaceType.COMBOBOX, msg)
-        return self._module.action(Selector.Action.GET_ALL_NAMES_FROM_SELECTION,
-                                   Selector.create_value_container(element=element, msg=msg),
-                                   msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.COMBOBOX, msg)
+        return module.action(Selector.Action.GET_ALL_NAMES_FROM_SELECTION,
+                             Selector.create_value_container(element=element, msg=msg),
+                             msg)
 
     @keyword
     def select_combobox_item_by_index(self, identifier, index, msg=None):
@@ -121,13 +125,14 @@ class ComboBoxKeywords:
         | Select Combobox Item By Index  <XPATH>  <INDEX> |
 
         """
-        element = self._module.get_element(identifier, InterfaceType.COMBOBOX, msg)
-        self._module.action(Selector.Action.SELECT_ITEM_BY_INDEX,
-                            Selector.create_value_container(element=element, index=index, msg=msg),
-                            msg)
-        self._module.action(Combobox.Action.COLLAPSE_COMBOBOX,
-                            Selector.create_value_container(element=element),
-                            msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.COMBOBOX, msg)
+        module.action(Selector.Action.SELECT_ITEM_BY_INDEX,
+                      Selector.create_value_container(element=element, index=index, msg=msg),
+                      msg)
+        module.action(Combobox.Action.COLLAPSE_COMBOBOX,
+                      Selector.create_value_container(element=element),
+                      msg)
 
     @keyword
     def combobox_should_contain(self, identifier, name, msg=None):
@@ -148,10 +153,11 @@ class ComboBoxKeywords:
         | Combobox Should Contain  <XPATH>  <NAME> |
 
         """
-        element = self._module.get_element(identifier, InterfaceType.COMBOBOX, msg)
-        self._module.action(Selector.Action.SHOULD_CONTAIN,
-                            Selector.create_value_container(element=element, name=name, msg=msg),
-                            msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.COMBOBOX, msg)
+        module.action(Selector.Action.SHOULD_CONTAIN,
+                      Selector.create_value_container(element=element, name=name, msg=msg),
+                      msg)
 
     @keyword
     def get_combobox_items_count(self, identifier, msg=None):
@@ -172,10 +178,11 @@ class ComboBoxKeywords:
         | Should Be Equal  ${value}  ${COUNT}         |
 
         """
-        element = self._module.get_element(identifier, InterfaceType.COMBOBOX, msg)
-        return self._module.action(Selector.Action.GET_ITEMS_COUNT,
-                                   Selector.create_value_container(element=element, msg=msg),
-                                   msg=msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.COMBOBOX, msg)
+        return module.action(Selector.Action.GET_ITEMS_COUNT,
+                             Selector.create_value_container(element=element, msg=msg),
+                             msg=msg)
 
     @keyword
     def collapse_combobox(self, identifier, msg=None):
@@ -195,10 +202,11 @@ class ComboBoxKeywords:
         | Collapse Combobox  <XPATH> |
 
         """
-        element = self._module.get_element(identifier, InterfaceType.COMBOBOX, msg)
-        self._module.action(Combobox.Action.COLLAPSE_COMBOBOX,
-                            Selector.create_value_container(element=element),
-                            msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.COMBOBOX, msg)
+        module.action(Combobox.Action.COLLAPSE_COMBOBOX,
+                      Selector.create_value_container(element=element),
+                      msg)
 
     @keyword
     def expand_combobox(self, identifier, msg=None):
@@ -218,7 +226,8 @@ class ComboBoxKeywords:
         | Expand Combobox  <XPATH> |
 
         """
-        element = self._module.get_element(identifier, InterfaceType.COMBOBOX, msg)
-        self._module.action(Combobox.Action.EXPAND_COMBOBOX,
-                            Selector.create_value_container(element=element),
-                            msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.COMBOBOX, msg)
+        module.action(Combobox.Action.EXPAND_COMBOBOX,
+                      Selector.create_value_container(element=element),
+                      msg)