+++ src/FlaUILibrary/keywords/textbox.py
@@ -1,7 +1,7 @@
 from robotlibcore import keyword
 from FlaUILibrary.flaui.enum import InterfaceType
 from FlaUILibrary.flaui.module import Textbox
-from FlaUILibrary.flaui.automation.uia import UIA
+from FlaUILibrary.flaui.util.automationinterfacecontainer import AutomationInterfaceContainer
 
 
 class TextBoxKeywords:
@@ -9,13 +9,13 @@ class TextBoxKeywords:
     Interface implementation from robotframework usage for textbox keywords.
     """
 
-    def __init__(self, module: UIA):
+    def __init__(self, container: AutomationInterfaceContainer):
         """
         Constructor for textbox keywords.
 
-        ``module`` Automation framework module like UIA3 to handle element interaction.
+        ``container`` User automation container to handle element interaction.
         """
-        self._module = module
+        self._container = container
 
     @keyword
     def get_text_from_textbox(self, identifier, msg=None):
@@ -38,10 +38,11 @@ class TextBoxKeywords:
         | Text string from textbox |
 
         """
-        element = self._module.get_element(identifier, InterfaceType.TEXTBOX, msg=msg)
-        return self._module.action(Textbox.Action.GET_TEXT_FROM_TEXTBOX,
-                                   Textbox.create_value_container(element=element),
-                                   msg=msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.TEXTBOX, msg=msg)
+        return module.action(Textbox.Action.GET_TEXT_FROM_TEXTBOX,
+                             Textbox.create_value_container(element=element),
+                             msg=msg)
 
     @keyword
     def set_text_to_textbox(self, identifier, value, msg=None):
@@ -62,7 +63,8 @@ class TextBoxKeywords:
         | Set Text To Textbox  <XPATH>  <VALUE> |
 
         """
-        element = self._module.get_element(identifier, InterfaceType.TEXTBOX, msg=msg)
-        self._module.action(Textbox.Action.SET_TEXT_TO_TEXTBOX,
-                            Textbox.create_value_container(element=element, value=value),
-                            msg)
+        module = self._container.get_module()
+        element = module.get_element(identifier, InterfaceType.TEXTBOX, msg=msg)
+        module.action(Textbox.Action.SET_TEXT_TO_TEXTBOX,
+                      Textbox.create_value_container(element=element, value=value),
+                      msg)